Initial commit
This is an Honors project that I did in my Java 2 class. The project fixes a problem that I actually had while working in the kitchen. They still do printer systems for tickets of food for each station. My thought was to create an application that could be run and output the tickets on a computer screen in a nice clear format.